+++ test/TensorValuesTests/OperationsTests.jl
@@ -355,7 +355,7 @@ v = TensorValue{2,3}(1,0,0,1,0,0)
 
 v = TensorValue{2,3}(1,0,0,1,1,0)
 @test meas(v) ≈ sqrt(2)
- 
+
 # Broadcasted operations
 
 a = VectorValue(1,2,3)
@@ -503,7 +503,7 @@ Sym4TensorIndexing = [1111, 1121, 1131, 1122, 1132, 1133, 2111, 2121, 2131, 2122
                       3111, 3121, 3131, 3122, 3132, 3133, 2211, 2221, 2231, 2222, 2232, 2233,
                       2311, 2321, 2331, 2322, 2332, 2333, 3311, 3321, 3331, 3322, 3332, 3333]
 test1 = test2 = SymFourthOrderTensorValue(1:36...)
-result = Int64[]
+result = Int[]
 for off_index in Sym4TensorIndexing
   i = parse(Int,string(off_index)[1]); j = parse(Int,string(off_index)[2]);
   m = parse(Int,string(off_index)[3]); p = parse(Int,string(off_index)[4]);
